requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1
python-dotenv==1.0.0
atlassian-python-api==3.41.0
bcrypt==4.1.2
